refactor postcss components to separate file
