+++ src/index.ts
@@ -19,13 +19,14 @@ export * from './mergify';
 export * from './next';
 export * from './node-project';
 export * from './project';
+export * from './pr-template';
 export * from './react';
+export * from './sample-file';
 export * from './semver';
 export * from './start';
+export * from './textfile';
 export * from './toml';
 export * from './typescript';
 export * from './version';
-export * from './textfile';
-export * from './pr-template';
-export * from './sample-file';
+export * from './web';
 export * from './yaml';
